[LIS-17] Updated module-base testing framework. Added tests for bundles and exceptions. CS Fixer
