+++ Test/Extra/DiscountHelperInterfaceFactory.php
@@ -8,8 +8,6 @@
 
 namespace Mygento\Base\Test\Extra;
 
-require_once __DIR__ . '/../../vendor/generate/generated/code/Mygento/Base/Api/DiscountHelperInterfaceFactory.php';
-
 /**
  * Factory class for @see \Mygento\Base\Api\DiscountHelperInterface
  */
