+++ Test/Unit/DiscountHelper/SplitItemsTest.php
@@ -6,14 +6,16 @@
  * @package Mygento_Base
  */
 
-namespace Mygento\Base\Test\Unit;
+namespace Mygento\Base\Test\Unit\DiscountHelper;
 
-class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
+use Mygento\Base\Api\DiscountHelperInterface;
+use Mygento\Base\Test\OrderMockBuilder;
+
+class SplitItemsTest extends GeneralTestCase
 {
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->discountHelper = $this->getDiscountHelperInstance();
-        $this->discountHelper->setSpreadDiscOnAllUnits(true);
         $this->discountHelper->setIsSplitItemsAllowed(true);
     }
 
@@ -46,6 +48,100 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
         }
     }
 
+    /** Test splitting item mechanism
+     *
+     * @dataProvider dataProviderItemsForSplitting
+     * @param mixed $item
+     * @param mixed $expectedArray
+     */
+    public function testProcessedItem($item, $expectedArray)
+    {
+        $discountHelper = $this->getDiscountHelperInstance();
+        $discountHelper->setIsSplitItemsAllowed(true);
+
+        $dHelper = new \ReflectionClass($discountHelper);
+        $getProcessedItem = $dHelper->getMethod('getProcessedItem');
+        $getProcessedItem->setAccessible(true);
+
+        $split = $getProcessedItem->invoke($discountHelper, $item);
+
+        $this->assertEquals(count($split), count($expectedArray), 'Item was not splitted correctly!');
+
+        $i = 0;
+        foreach ($split as $item) {
+            $this->assertEquals($expectedArray[$i]['price'], $item['price'], 'Price of item failed');
+            $this->assertEquals($expectedArray[$i]['quantity'], $item['quantity']);
+            $this->assertEquals($expectedArray[$i]['sum'], $item['sum'], 'Sum of item failed');
+
+            $i++;
+        }
+    }
+
+    /**
+     * @return array
+     * @SuppressWarnings(PHPMD)
+     */
+    public function dataProviderItemsForSplitting()
+    {
+        $final = [];
+
+        // #1 rowDiff = 2 kop. qty = 3. qtyUpdate = 3
+        $item = OrderMockBuilder::getItem(0, 0, 0, 3);
+        $item->setData(DiscountHelperInterface::NAME_ROW_DIFF, 2);
+        $item->setData(DiscountHelperInterface::NAME_UNIT_PRICE, 10.59);
+
+        $expected = [
+            [
+                'price' => 10.59,
+                'quantity' => 1,
+                'sum' => 10.59,
+                'tax' => null,
+            ],
+            [
+                'price' => 10.6,
+                'quantity' => 2,
+                'sum' => 21.2,
+                'tax' => null,
+            ],
+        ];
+        $final['#case 1. 2 копейки распределить по 3м товарам.'] = [$item, $expected];
+
+        // #2 rowDiff = 150 kop. qty = 30. qtyUpdate = 0
+        $item2 = OrderMockBuilder::getItem(0, 0, 0, 30);
+        $item2->setData(DiscountHelperInterface::NAME_ROW_DIFF, 150);
+        $item2->setData(DiscountHelperInterface::NAME_UNIT_PRICE, 10);
+
+        $expected2 = [
+            [
+                'price' => 10.05,
+                'quantity' => 30,
+                'sum' => 301.5,
+            ],
+        ];
+        $final['#case 2. 150 копеек распределить по 30 товарам.'] = [$item2, $expected2];
+
+        // #3 rowDiff = 5 kop. qty = 3. qtyUpdate = 2
+        $item3 = OrderMockBuilder::getItem(0, 0, 0, 3);
+        $item3->setData(DiscountHelperInterface::NAME_ROW_DIFF, 5);
+        $item3->setData(DiscountHelperInterface::NAME_UNIT_PRICE, 10);
+
+        $expected3 = [
+            [
+                'price' => 10.01,
+                'quantity' => 1,
+                'sum' => 10.01,
+            ],
+            [
+                'price' => 10.02,
+                'quantity' => 2,
+                'sum' => 20.04,
+            ],
+        ];
+        $final['#case 3. 5 копеек распределить по 3 товарам.'] = [$item3, $expected3];
+
+        return $final;
+    }
+
     /**
      * @SuppressWarnings(PHPMD.ExcessiveMethodLength)
      */
@@ -56,14 +152,14 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'origGrandTotal' => 12069.30,
             'items' => [
                 152 => [
-                    'price' => 11717.5,
+                    'price' => 11691,
                     'quantity' => 1,
-                    'sum' => 11717.5,
+                    'sum' => 11691,
                 ],
                 153 => [
-                    'price' => 351.8,
+                    'price' => 378.30,
                     'quantity' => 1,
-                    'sum' => 351.8,
+                    'sum' => 378.30,
                 ],
                 154 => [
                     'price' => 0,
@@ -83,19 +179,19 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'origGrandTotal' => 9373.19,
             'items' => [
                 152 => [
-                    'price' => 5015.28,
+                    'price' => 5054.4,
                     'quantity' => 1,
-                    'sum' => 5015.28,
+                    'sum' => 5054.4,
                 ],
-                '153_1' => [
-                    'price' => 23.63,
+                '154_1' => [
+                    'price' => 10.59,
                     'quantity' => 1,
-                    'sum' => 23.63,
+                    'sum' => 10.59,
                 ],
                 '153_2' => [
-                    'price' => 23.64,
+                    'price' => 10.6,
                     'quantity' => 2,
-                    'sum' => 47.28,
+                    'sum' => 21.2,
                 ],
                 'shipping' => [
                     'price' => 4287.00,
@@ -159,24 +255,24 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'origGrandTotal' => 202.1,
             'items' => [
                 '152_1' => [
-                    'price' => 36.41,
-                    'quantity' => 2,
-                    'sum' => 72.82,
+                    'price' => 33.66,
+                    'quantity' => 1,
+                    'sum' => 33.66,
                 ],
                 '152_2' => [
-                    'price' => 36.42,
-                    'quantity' => 1,
-                    'sum' => 36.42,
+                    'price' => 33.67,
+                    'quantity' => 2,
+                    'sum' => 67.34,
                 ],
                 '153_1' => [
-                    'price' => 23.21,
+                    'price' => 25.27,
                     'quantity' => 2,
-                    'sum' => 46.42,
+                    'sum' => 50.54,
                 ],
                 '153_2' => [
-                    'price' => 23.22,
+                    'price' => 25.28,
                     'quantity' => 2,
-                    'sum' => 46.44,
+                    'sum' => 50.56,
                 ],
                 'shipping' => [
                     'price' => 0.0,
@@ -191,34 +287,29 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'origGrandTotal' => 702.1,
             'items' => [
                 '152_1' => [
-                    'price' => 38.89,
+                    'price' => 33.66,
                     'quantity' => 1,
-                    'sum' => 38.89,
+                    'sum' => 33.66,
                 ],
                 '152_2' => [
-                    'price' => 38.9,
+                    'price' => 33.67,
                     'quantity' => 2,
-                    'sum' => 77.8,
+                    'sum' => 67.34,
                 ],
                 '153_1' => [
-                    'price' => 24.79,
-                    'quantity' => 1,
-                    'sum' => 24.79,
+                    'price' => 25.27,
+                    'quantity' => 2,
+                    'sum' => 50.54,
                 ],
                 '153_2' => [
-                    'price' => 24.8,
-                    'quantity' => 3,
-                    'sum' => 74.40,
-                ],
-                '154_1' => [
-                    'price' => 97.24,
-                    'quantity' => 3,
-                    'sum' => 291.72,
-                ],
-                '154_2' => [
-                    'price' => 97.25,
+                    'price' => 25.28,
                     'quantity' => 2,
-                    'sum' => 194.5,
+                    'sum' => 50.56,
+                ],
+                154 => [
+                    'price' => 100,
+                    'quantity' => 5,
+                    'sum' => 500,
                 ],
                 'shipping' => [
                     'price' => 0.00,
@@ -233,14 +324,14 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'origGrandTotal' => 11691.0,
             'items' => [
                 152 => [
-                    'price' => 11673.03,
+                    'price' => 11691.0,
                     'quantity' => 1,
-                    'sum' => 11673.03,
+                    'sum' => 11691.0,
                 ],
                 153 => [
-                    'price' => 17.97,
+                    'price' => 0.0,
                     'quantity' => 1,
-                    'sum' => 17.97,
+                    'sum' => 0.0,
                 ],
                 'shipping' => [
                     'price' => 0.00,
@@ -255,14 +346,14 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'origGrandTotal' => 11611.0,
             'items' => [
                 152 => [
-                    'price' => 11593.15,
+                    'price' => 11591.15,
                     'quantity' => 1,
-                    'sum' => 11593.15,
+                    'sum' => 11591.15,
                 ],
                 153 => [
-                    'price' => 17.85,
+                    'price' => 19.85,
                     'quantity' => 1,
-                    'sum' => 17.85,
+                    'sum' => 19.85,
                 ],
                 'shipping' => [
                     'price' => 0.00,
@@ -273,22 +364,18 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
         ];
 
         $actualData[parent::TEST_CASE_NAME_9] = [
-            'sum' => 12890,
-            'origGrandTotal' => 12890,
+            'sum' => 12890.0,
+            'origGrandTotal' => 12890.0,
             'items' => [
-                0 => [
-                    'price' => 12890,
-                    'name' => 'qGsDb5jK',
+                152 => [
+                    'price' => 12890.0,
                     'quantity' => 1,
-                    'sum' => 12890,
-                    'tax' => 'vat20',
+                    'sum' => 12890.0,
                 ],
                 'shipping' => [
-                    'name' => '',
-                    'price' => 0,
+                    'price' => 0.00,
                     'quantity' => 1,
-                    'sum' => 0,
-                    'tax' => '',
+                    'sum' => 0.00,
                 ],
             ],
         ];
@@ -320,49 +407,30 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'origGrandTotal' => 32130.01,
             'items' => [
                 0 => [
-                    'price' => 17298.11,
-                    'name' => 'tVPQqM6P',
+                    'price' => 19990,
                     'quantity' => 1,
-                    'sum' => 17298.11,
-                    'tax' => 'vat20',
-                ],
-                '100523_1' => [
-                    'price' => 25.09,
-                    'name' => 'dawEIFWS',
-                    'quantity' => 260,
-                    'sum' => 6523.4,
-                    'tax' => 'vat20',
-                ],
-                '100523_2' => [
-                    'price' => 25.1,
-                    'name' => 'dawEIFWS',
-                    'quantity' => 240,
-                    'sum' => 6024,
+                    'sum' => 19990,
                     'tax' => 'vat20',
                 ],
                 1 => [
-                    'price' => 865.35,
-                    'name' => 'GLr5lhBO',
-                    'quantity' => 1,
-                    'sum' => 865.35,
+                    'price' => 19,
+                    'quantity' => 500,
+                    'sum' => 9500,
                     'tax' => 'vat20',
                 ],
-                '100525_1' => [
-                    'price' => 354.78,
-                    'name' => 'DgCaf1zm',
+                2 => [
+                    'price' => 1000.01,
                     'quantity' => 1,
-                    'sum' => 354.78,
+                    'sum' => 1000.01,
                     'tax' => 'vat20',
                 ],
-                '100525_2' => [
-                    'price' => 354.79,
-                    'name' => 'DgCaf1zm',
-                    'quantity' => 3,
-                    'sum' => 1064.37,
+                3 => [
+                    'price' => 410,
+                    'quantity' => 4,
+                    'sum' => 1640,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
-                    'name' => '',
                     'price' => 0,
                     'quantity' => 1,
                     'sum' => 0,
@@ -376,143 +444,150 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'origGrandTotal' => 13189.99,
             'items' => [
                 0 => [
-                    'price' => 5793.74,
-                    'name' => 'KJmABDf6',
+                    'price' => 7989.99,
+                    'name' => 'HbcIyFpc',
                     'quantity' => 1,
-                    'sum' => 5793.74,
+                    'sum' => 7989.99,
                     'tax' => 'vat20',
                 ],
                 '100527_1' => [
-                    'price' => 26.1,
-                    'name' => '6nFZC8V8',
-                    'quantity' => 22,
-                    'sum' => 574.2,
+                    'price' => 18.35,
+                    'name' => 'iVZQ2iMO',
+                    'quantity' => 28,
+                    'sum' => 513.8,
                     'tax' => 'vat20',
                 ],
                 '100527_2' => [
-                    'price' => 26.11,
-                    'name' => '6nFZC8V8',
-                    'quantity' => 18,
-                    'sum' => 469.98,
+                    'price' => 18.36,
+                    'name' => 'iVZQ2iMO',
+                    'quantity' => 12,
+                    'sum' => 220.32,
                     'tax' => 'vat20',
                 ],
                 '100528_1' => [
-                    'price' => 26.1,
-                    'name' => '0YMYJeK7',
-                    'quantity' => 17,
-                    'sum' => 443.7,
+                    'price' => 18.35,
+                    'name' => 'CVUohjpK',
+                    'quantity' => 21,
+                    'sum' => 385.35,
                     'tax' => 'vat20',
                 ],
                 '100528_2' => [
-                    'price' => 26.11,
-                    'name' => '0YMYJeK7',
-                    'quantity' => 13,
-                    'sum' => 339.43,
+                    'price' => 18.36,
+                    'name' => 'CVUohjpK',
+                    'quantity' => 9,
+                    'sum' => 165.24,
                     'tax' => 'vat20',
                 ],
                 '100529_1' => [
-                    'price' => 21.02,
-                    'name' => '1jPzAF9j',
-                    'quantity' => 6,
-                    'sum' => 126.12,
+                    'price' => 14.78,
+                    'name' => '3JFWNpUY',
+                    'quantity' => 23,
+                    'sum' => 339.94,
                     'tax' => 'vat20',
                 ],
                 '100529_2' => [
-                    'price' => 21.03,
-                    'name' => '1jPzAF9j',
-                    'quantity' => 34,
-                    'sum' => 715.02,
+                    'price' => 14.79,
+                    'name' => '3JFWNpUY',
+                    'quantity' => 17,
+                    'sum' => 251.43,
                     'tax' => 'vat20',
                 ],
                 '100530_1' => [
-                    'price' => 21.02,
-                    'name' => 'gD8Tjaok',
-                    'quantity' => 7,
-                    'sum' => 147.14,
+                    'price' => 14.78,
+                    'name' => 'eLjly6un',
+                    'quantity' => 28,
+                    'sum' => 413.84,
                     'tax' => 'vat20',
                 ],
                 '100530_2' => [
-                    'price' => 21.03,
-                    'name' => 'gD8Tjaok',
-                    'quantity' => 43,
-                    'sum' => 904.29,
+                    'price' => 14.79,
+                    'name' => 'eLjly6un',
+                    'quantity' => 22,
+                    'sum' => 325.38,
                     'tax' => 'vat20',
                 ],
                 '100531_1' => [
-                    'price' => 26.1,
-                    'name' => 'rLfMKyoC',
-                    'quantity' => 17,
-                    'sum' => 443.7,
+                    'price' => 18.35,
+                    'name' => 'nZ0KslHN',
+                    'quantity' => 21,
+                    'sum' => 385.35,
                     'tax' => 'vat20',
                 ],
                 '100531_2' => [
-                    'price' => 26.11,
-                    'name' => 'rLfMKyoC',
-                    'quantity' => 13,
-                    'sum' => 339.43,
+                    'price' => 18.36,
+                    'name' => 'nZ0KslHN',
+                    'quantity' => 9,
+                    'sum' => 165.24,
                     'tax' => 'vat20',
                 ],
                 '100532_1' => [
-                    'price' => 26.1,
-                    'name' => 'L4IqEshr',
-                    'quantity' => 6,
-                    'sum' => 156.6,
+                    'price' => 18.35,
+                    'name' => 'xk0eWAiC',
+                    'quantity' => 7,
+                    'sum' => 128.45,
                     'tax' => 'vat20',
                 ],
                 '100532_2' => [
-                    'price' => 26.11,
-                    'name' => 'L4IqEshr',
-                    'quantity' => 4,
-                    'sum' => 104.44,
+                    'price' => 18.36,
+                    'name' => 'xk0eWAiC',
+                    'quantity' => 3,
+                    'sum' => 55.08,
                     'tax' => 'vat20',
                 ],
                 '100533_1' => [
-                    'price' => 26.1,
-                    'name' => 'l4p1n791',
-                    'quantity' => 28,
-                    'sum' => 730.8,
+                    'price' => 18.35,
+                    'name' => 'QZc84oGJ',
+                    'quantity' => 35,
+                    'sum' => 642.25,
                     'tax' => 'vat20',
                 ],
                 '100533_2' => [
-                    'price' => 26.11,
-                    'name' => 'l4p1n791',
-                    'quantity' => 22,
-                    'sum' => 574.42,
+                    'price' => 18.36,
+                    'name' => 'QZc84oGJ',
+                    'quantity' => 15,
+                    'sum' => 275.4,
                     'tax' => 'vat20',
                 ],
                 '100534_1' => [
-                    'price' => 23.92,
-                    'name' => 'hfzoDuFf',
-                    'quantity' => 1,
-                    'sum' => 23.92,
+                    'price' => 16.82,
+                    'name' => 'EZ45M8YX',
+                    'quantity' => 6,
+                    'sum' => 100.92,
                     'tax' => 'vat20',
                 ],
                 '100534_2' => [
-                    'price' => 23.93,
-                    'name' => 'hfzoDuFf',
-                    'quantity' => 9,
-                    'sum' => 215.37,
+                    'price' => 16.83,
+                    'name' => 'EZ45M8YX',
+                    'quantity' => 4,
+                    'sum' => 67.32,
                     'tax' => 'vat20',
                 ],
                 '100535_1' => [
-                    'price' => 26.1,
-                    'name' => '8PWR3Pop',
-                    'quantity' => 11,
-                    'sum' => 287.1,
+                    'price' => 18.35,
+                    'name' => '1fSGTfUL',
+                    'quantity' => 14,
+                    'sum' => 256.9,
                     'tax' => 'vat20',
                 ],
                 '100535_2' => [
-                    'price' => 26.11,
-                    'name' => '8PWR3Pop',
-                    'quantity' => 9,
-                    'sum' => 234.99,
+                    'price' => 18.36,
+                    'name' => '1fSGTfUL',
+                    'quantity' => 6,
+                    'sum' => 110.16,
                     'tax' => 'vat20',
                 ],
-                1 => [
-                    'price' => 28.28,
-                    'name' => 'kOto9Vfv',
-                    'quantity' => 20,
-                    'sum' => 565.6,
+                '100536_1' => [
+                    'price' => 19.88,
+                    'name' => 'KK1Iub5Q',
+                    'quantity' => 17,
+                    'sum' => 337.96,
+                    'tax' => 'vat20',
+                ],
+                '100536_2' => [
+                    'price' => 19.89,
+                    'name' => 'KK1Iub5Q',
+                    'quantity' => 3,
+                    'sum' => 59.67,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
@@ -531,142 +606,142 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'items' => [
                 '100537_1' => [
                     'price' => 18.35,
-                    'name' => 'LAgOo1AI',
-                    'quantity' => 28,
-                    'sum' => 513.8,
+                    'name' => 'QasOwBxx',
+                    'quantity' => 29,
+                    'sum' => 532.15,
                     'tax' => 'vat20',
                 ],
                 '100537_2' => [
                     'price' => 18.36,
-                    'name' => 'LAgOo1AI',
-                    'quantity' => 12,
-                    'sum' => 220.32,
+                    'name' => 'QasOwBxx',
+                    'quantity' => 11,
+                    'sum' => 201.96,
                     'tax' => 'vat20',
                 ],
                 '100538_1' => [
                     'price' => 18.35,
-                    'name' => 'NqIQuqd7',
+                    'name' => 'i3q2Pqi2',
                     'quantity' => 21,
                     'sum' => 385.35,
                     'tax' => 'vat20',
                 ],
                 '100538_2' => [
                     'price' => 18.36,
-                    'name' => 'NqIQuqd7',
+                    'name' => 'i3q2Pqi2',
                     'quantity' => 9,
                     'sum' => 165.24,
                     'tax' => 'vat20',
                 ],
                 '100539_1' => [
                     'price' => 14.78,
-                    'name' => 'xzZITJyE',
+                    'name' => '4yAnOQ9Q',
                     'quantity' => 23,
                     'sum' => 339.94,
                     'tax' => 'vat20',
                 ],
                 '100539_2' => [
                     'price' => 14.79,
-                    'name' => 'xzZITJyE',
+                    'name' => '4yAnOQ9Q',
                     'quantity' => 17,
                     'sum' => 251.43,
                     'tax' => 'vat20',
                 ],
                 '100540_1' => [
                     'price' => 14.78,
-                    'name' => 'wKiM556l',
-                    'quantity' => 29,
-                    'sum' => 428.62,
+                    'name' => 'BpzRVt8O',
+                    'quantity' => 28,
+                    'sum' => 413.84,
                     'tax' => 'vat20',
                 ],
                 '100540_2' => [
                     'price' => 14.79,
-                    'name' => 'wKiM556l',
-                    'quantity' => 21,
-                    'sum' => 310.59,
+                    'name' => 'BpzRVt8O',
+                    'quantity' => 22,
+                    'sum' => 325.38,
                     'tax' => 'vat20',
                 ],
                 '100541_1' => [
                     'price' => 18.35,
-                    'name' => 'n1MkSUN8',
+                    'name' => 'bOsve0El',
                     'quantity' => 21,
                     'sum' => 385.35,
                     'tax' => 'vat20',
                 ],
                 '100541_2' => [
                     'price' => 18.36,
-                    'name' => 'n1MkSUN8',
+                    'name' => 'bOsve0El',
                     'quantity' => 9,
                     'sum' => 165.24,
                     'tax' => 'vat20',
                 ],
                 '100542_1' => [
                     'price' => 18.35,
-                    'name' => 'yG0B7EKY',
+                    'name' => 'VD4fh7ow',
                     'quantity' => 7,
                     'sum' => 128.45,
                     'tax' => 'vat20',
                 ],
                 '100542_2' => [
                     'price' => 18.36,
-                    'name' => 'yG0B7EKY',
+                    'name' => 'VD4fh7ow',
                     'quantity' => 3,
                     'sum' => 55.08,
                     'tax' => 'vat20',
                 ],
                 '100543_1' => [
                     'price' => 18.35,
-                    'name' => 'P2aMixwX',
-                    'quantity' => 36,
-                    'sum' => 660.6,
+                    'name' => '7rUyXogC',
+                    'quantity' => 35,
+                    'sum' => 642.25,
                     'tax' => 'vat20',
                 ],
                 '100543_2' => [
                     'price' => 18.36,
-                    'name' => 'P2aMixwX',
-                    'quantity' => 14,
-                    'sum' => 257.04,
+                    'name' => '7rUyXogC',
+                    'quantity' => 15,
+                    'sum' => 275.4,
                     'tax' => 'vat20',
                 ],
                 '100544_1' => [
                     'price' => 16.82,
-                    'name' => 'RdIoXoWg',
-                    'quantity' => 7,
-                    'sum' => 117.74,
+                    'name' => '4Vjv1sDw',
+                    'quantity' => 6,
+                    'sum' => 100.92,
                     'tax' => 'vat20',
                 ],
                 '100544_2' => [
                     'price' => 16.83,
-                    'name' => 'RdIoXoWg',
-                    'quantity' => 3,
-                    'sum' => 50.49,
+                    'name' => '4Vjv1sDw',
+                    'quantity' => 4,
+                    'sum' => 67.32,
                     'tax' => 'vat20',
                 ],
                 '100545_1' => [
                     'price' => 18.35,
-                    'name' => 'i0mX1Des',
+                    'name' => 'I1v7ozqi',
                     'quantity' => 14,
                     'sum' => 256.9,
                     'tax' => 'vat20',
                 ],
                 '100545_2' => [
                     'price' => 18.36,
-                    'name' => 'i0mX1Des',
+                    'name' => 'I1v7ozqi',
                     'quantity' => 6,
                     'sum' => 110.16,
                     'tax' => 'vat20',
                 ],
                 '100546_1' => [
                     'price' => 19.88,
-                    'name' => 'u40iqwaA',
-                    'quantity' => 15,
-                    'sum' => 298.2,
+                    'name' => 'vPZr7cV2',
+                    'quantity' => 17,
+                    'sum' => 337.96,
                     'tax' => 'vat20',
                 ],
                 '100546_2' => [
                     'price' => 19.89,
-                    'name' => 'u40iqwaA',
-                    'quantity' => 5,
-                    'sum' => 99.45,
+                    'name' => 'vPZr7cV2',
+                    'quantity' => 3,
+                    'sum' => 59.67,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
@@ -684,143 +759,150 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'origGrandTotal' => 13190.01,
             'items' => [
                 0 => [
-                    'price' => 5793.75,
-                    'name' => '84MEoc1Y',
+                    'price' => 7990.01,
+                    'name' => 'BPIXObQh',
                     'quantity' => 1,
-                    'sum' => 5793.75,
+                    'sum' => 7990.01,
                     'tax' => 'vat20',
                 ],
                 '100548_1' => [
-                    'price' => 26.1,
-                    'name' => 'VDAYw0zS',
-                    'quantity' => 21,
-                    'sum' => 548.1,
+                    'price' => 18.35,
+                    'name' => 'kXtzFQk9',
+                    'quantity' => 28,
+                    'sum' => 513.8,
                     'tax' => 'vat20',
                 ],
                 '100548_2' => [
-                    'price' => 26.11,
-                    'name' => 'VDAYw0zS',
-                    'quantity' => 19,
-                    'sum' => 496.09,
+                    'price' => 18.36,
+                    'name' => 'kXtzFQk9',
+                    'quantity' => 12,
+                    'sum' => 220.32,
                     'tax' => 'vat20',
                 ],
                 '100549_1' => [
-                    'price' => 26.1,
-                    'name' => '4UUBllTv',
-                    'quantity' => 17,
-                    'sum' => 443.7,
+                    'price' => 18.35,
+                    'name' => 'VjBZWKqC',
+                    'quantity' => 21,
+                    'sum' => 385.35,
                     'tax' => 'vat20',
                 ],
                 '100549_2' => [
-                    'price' => 26.11,
-                    'name' => '4UUBllTv',
-                    'quantity' => 13,
-                    'sum' => 339.43,
+                    'price' => 18.36,
+                    'name' => 'VjBZWKqC',
+                    'quantity' => 9,
+                    'sum' => 165.24,
                     'tax' => 'vat20',
                 ],
                 '100550_1' => [
-                    'price' => 21.02,
-                    'name' => 'lVP8ceEm',
-                    'quantity' => 6,
-                    'sum' => 126.12,
+                    'price' => 14.78,
+                    'name' => 'yOhCey5i',
+                    'quantity' => 23,
+                    'sum' => 339.94,
                     'tax' => 'vat20',
                 ],
                 '100550_2' => [
-                    'price' => 21.03,
-                    'name' => 'lVP8ceEm',
-                    'quantity' => 34,
-                    'sum' => 715.02,
+                    'price' => 14.79,
+                    'name' => 'yOhCey5i',
+                    'quantity' => 17,
+                    'sum' => 251.43,
                     'tax' => 'vat20',
                 ],
                 '100551_1' => [
-                    'price' => 21.02,
-                    'name' => 'Lx52oORB',
-                    'quantity' => 7,
-                    'sum' => 147.14,
+                    'price' => 14.78,
+                    'name' => 'x744Y2VU',
+                    'quantity' => 28,
+                    'sum' => 413.84,
                     'tax' => 'vat20',
                 ],
                 '100551_2' => [
-                    'price' => 21.03,
-                    'name' => 'Lx52oORB',
-                    'quantity' => 43,
-                    'sum' => 904.29,
+                    'price' => 14.79,
+                    'name' => 'x744Y2VU',
+                    'quantity' => 22,
+                    'sum' => 325.38,
                     'tax' => 'vat20',
                 ],
                 '100552_1' => [
-                    'price' => 26.1,
-                    'name' => '9APZJHJX',
-                    'quantity' => 17,
-                    'sum' => 443.7,
+                    'price' => 18.35,
+                    'name' => 'LNgNKsqq',
+                    'quantity' => 21,
+                    'sum' => 385.35,
                     'tax' => 'vat20',
                 ],
                 '100552_2' => [
-                    'price' => 26.11,
-                    'name' => '9APZJHJX',
-                    'quantity' => 13,
-                    'sum' => 339.43,
+                    'price' => 18.36,
+                    'name' => 'LNgNKsqq',
+                    'quantity' => 9,
+                    'sum' => 165.24,
                     'tax' => 'vat20',
                 ],
                 '100553_1' => [
-                    'price' => 26.1,
-                    'name' => '1jSEl731',
-                    'quantity' => 6,
-                    'sum' => 156.6,
+                    'price' => 18.35,
+                    'name' => '4vrYuAmx',
+                    'quantity' => 7,
+                    'sum' => 128.45,
                     'tax' => 'vat20',
                 ],
                 '100553_2' => [
-                    'price' => 26.11,
-                    'name' => '1jSEl731',
-                    'quantity' => 4,
-                    'sum' => 104.44,
+                    'price' => 18.36,
+                    'name' => '4vrYuAmx',
+                    'quantity' => 3,
+                    'sum' => 55.08,
                     'tax' => 'vat20',
                 ],
                 '100554_1' => [
-                    'price' => 26.1,
-                    'name' => 'mU5bfDEV',
-                    'quantity' => 28,
-                    'sum' => 730.8,
+                    'price' => 18.35,
+                    'name' => '5KCqwVCK',
+                    'quantity' => 35,
+                    'sum' => 642.25,
                     'tax' => 'vat20',
                 ],
                 '100554_2' => [
-                    'price' => 26.11,
-                    'name' => 'mU5bfDEV',
-                    'quantity' => 22,
-                    'sum' => 574.42,
+                    'price' => 18.36,
+                    'name' => '5KCqwVCK',
+                    'quantity' => 15,
+                    'sum' => 275.4,
                     'tax' => 'vat20',
                 ],
                 '100555_1' => [
-                    'price' => 23.92,
-                    'name' => '3seC8xNP',
-                    'quantity' => 1,
-                    'sum' => 23.92,
+                    'price' => 16.82,
+                    'name' => 'SiJ3Zm9y',
+                    'quantity' => 6,
+                    'sum' => 100.92,
                     'tax' => 'vat20',
                 ],
                 '100555_2' => [
-                    'price' => 23.93,
-                    'name' => '3seC8xNP',
-                    'quantity' => 9,
-                    'sum' => 215.37,
+                    'price' => 16.83,
+                    'name' => 'SiJ3Zm9y',
+                    'quantity' => 4,
+                    'sum' => 67.32,
                     'tax' => 'vat20',
                 ],
                 '100556_1' => [
-                    'price' => 26.1,
-                    'name' => 'Yf5U4J8g',
-                    'quantity' => 11,
-                    'sum' => 287.1,
+                    'price' => 18.35,
+                    'name' => '2IqCgAK6',
+                    'quantity' => 14,
+                    'sum' => 256.9,
                     'tax' => 'vat20',
                 ],
                 '100556_2' => [
-                    'price' => 26.11,
-                    'name' => 'Yf5U4J8g',
-                    'quantity' => 9,
-                    'sum' => 234.99,
+                    'price' => 18.36,
+                    'name' => '2IqCgAK6',
+                    'quantity' => 6,
+                    'sum' => 110.16,
                     'tax' => 'vat20',
                 ],
-                1 => [
-                    'price' => 28.28,
-                    'name' => 'cbn8w9ya',
-                    'quantity' => 20,
-                    'sum' => 565.6,
+                '100557_1' => [
+                    'price' => 19.88,
+                    'name' => 'UDzfCEuJ',
+                    'quantity' => 17,
+                    'sum' => 337.96,
+                    'tax' => 'vat20',
+                ],
+                '100557_2' => [
+                    'price' => 19.89,
+                    'name' => 'UDzfCEuJ',
+                    'quantity' => 3,
+                    'sum' => 59.67,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
@@ -838,150 +920,150 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'origGrandTotal' => 13189.69,
             'items' => [
                 0 => [
-                    'price' => 5793.6,
-                    'name' => 'kz5oHmL6',
+                    'price' => 7989.96,
+                    'name' => 'WK6xqA9P',
                     'quantity' => 1,
-                    'sum' => 5793.6,
+                    'sum' => 7989.96,
                     'tax' => 'vat20',
                 ],
                 '100559_1' => [
-                    'price' => 26.1,
-                    'name' => 'cewKmxkL',
-                    'quantity' => 25,
-                    'sum' => 652.5,
+                    'price' => 18.35,
+                    'name' => 'TGKkaZv4',
+                    'quantity' => 32,
+                    'sum' => 587.2,
                     'tax' => 'vat20',
                 ],
                 '100559_2' => [
-                    'price' => 26.11,
-                    'name' => 'cewKmxkL',
-                    'quantity' => 15,
-                    'sum' => 391.65,
+                    'price' => 18.36,
+                    'name' => 'TGKkaZv4',
+                    'quantity' => 8,
+                    'sum' => 146.88,
                     'tax' => 'vat20',
                 ],
                 '100560_1' => [
-                    'price' => 26.1,
-                    'name' => 'Mkr1KjKu',
-                    'quantity' => 18,
-                    'sum' => 469.8,
+                    'price' => 18.35,
+                    'name' => 'EixOXbqy',
+                    'quantity' => 25,
+                    'sum' => 458.75,
                     'tax' => 'vat20',
                 ],
                 '100560_2' => [
-                    'price' => 26.11,
-                    'name' => 'Mkr1KjKu',
-                    'quantity' => 12,
-                    'sum' => 313.32,
+                    'price' => 18.36,
+                    'name' => 'EixOXbqy',
+                    'quantity' => 5,
+                    'sum' => 91.8,
                     'tax' => 'vat20',
                 ],
                 '100561_1' => [
-                    'price' => 21.02,
-                    'name' => '2lqFDqfm',
-                    'quantity' => 8,
-                    'sum' => 168.16,
+                    'price' => 14.78,
+                    'name' => 'kCJIu3aM',
+                    'quantity' => 27,
+                    'sum' => 399.06,
                     'tax' => 'vat20',
                 ],
                 '100561_2' => [
-                    'price' => 21.03,
-                    'name' => '2lqFDqfm',
-                    'quantity' => 32,
-                    'sum' => 672.96,
+                    'price' => 14.79,
+                    'name' => 'kCJIu3aM',
+                    'quantity' => 13,
+                    'sum' => 192.27,
                     'tax' => 'vat20',
                 ],
                 '100562_1' => [
-                    'price' => 21.02,
-                    'name' => '91IG7jzc',
-                    'quantity' => 10,
-                    'sum' => 210.2,
+                    'price' => 14.78,
+                    'name' => 'MdGYRsVO',
+                    'quantity' => 31,
+                    'sum' => 458.18,
                     'tax' => 'vat20',
                 ],
                 '100562_2' => [
-                    'price' => 21.03,
-                    'name' => '91IG7jzc',
-                    'quantity' => 40,
-                    'sum' => 841.2,
+                    'price' => 14.79,
+                    'name' => 'MdGYRsVO',
+                    'quantity' => 19,
+                    'sum' => 281.01,
                     'tax' => 'vat20',
                 ],
                 '100563_1' => [
-                    'price' => 26.1,
-                    'name' => 'ICYKymBz',
-                    'quantity' => 18,
-                    'sum' => 469.8,
+                    'price' => 18.35,
+                    'name' => 'ttQ8ylbR',
+                    'quantity' => 23,
+                    'sum' => 422.05,
                     'tax' => 'vat20',
                 ],
                 '100563_2' => [
-                    'price' => 26.11,
-                    'name' => 'ICYKymBz',
-                    'quantity' => 12,
-                    'sum' => 313.32,
+                    'price' => 18.36,
+                    'name' => 'ttQ8ylbR',
+                    'quantity' => 7,
+                    'sum' => 128.52,
                     'tax' => 'vat20',
                 ],
                 '100564_1' => [
-                    'price' => 26.1,
-                    'name' => 'gjn12tUv',
-                    'quantity' => 6,
-                    'sum' => 156.6,
+                    'price' => 18.35,
+                    'name' => 'wWrAvSS9',
+                    'quantity' => 9,
+                    'sum' => 165.15,
                     'tax' => 'vat20',
                 ],
                 '100564_2' => [
-                    'price' => 26.11,
-                    'name' => 'gjn12tUv',
-                    'quantity' => 4,
-                    'sum' => 104.44,
+                    'price' => 18.36,
+                    'name' => 'wWrAvSS9',
+                    'quantity' => 1,
+                    'sum' => 18.36,
                     'tax' => 'vat20',
                 ],
                 '100565_1' => [
-                    'price' => 26.1,
-                    'name' => '25wUYkjA',
-                    'quantity' => 31,
-                    'sum' => 809.1,
+                    'price' => 18.35,
+                    'name' => 'v415Myym',
+                    'quantity' => 37,
+                    'sum' => 678.95,
                     'tax' => 'vat20',
                 ],
                 '100565_2' => [
-                    'price' => 26.11,
-                    'name' => '25wUYkjA',
-                    'quantity' => 19,
-                    'sum' => 496.09,
+                    'price' => 18.36,
+                    'name' => 'v415Myym',
+                    'quantity' => 13,
+                    'sum' => 238.68,
                     'tax' => 'vat20',
                 ],
                 '100566_1' => [
-                    'price' => 23.92,
-                    'name' => 'iFEcrv17',
-                    'quantity' => 1,
-                    'sum' => 23.92,
+                    'price' => 16.82,
+                    'name' => '2KkA1lAQ',
+                    'quantity' => 8,
+                    'sum' => 134.56,
                     'tax' => 'vat20',
                 ],
                 '100566_2' => [
-                    'price' => 23.93,
-                    'name' => 'iFEcrv17',
-                    'quantity' => 9,
-                    'sum' => 215.37,
+                    'price' => 16.83,
+                    'name' => '2KkA1lAQ',
+                    'quantity' => 2,
+                    'sum' => 33.66,
                     'tax' => 'vat20',
                 ],
                 '100567_1' => [
-                    'price' => 26.1,
-                    'name' => 'znVtmS3R',
-                    'quantity' => 12,
-                    'sum' => 313.2,
+                    'price' => 18.35,
+                    'name' => 'Mbg5nGZU',
+                    'quantity' => 16,
+                    'sum' => 293.6,
                     'tax' => 'vat20',
                 ],
                 '100567_2' => [
-                    'price' => 26.11,
-                    'name' => 'znVtmS3R',
-                    'quantity' => 8,
-                    'sum' => 208.88,
+                    'price' => 18.36,
+                    'name' => 'Mbg5nGZU',
+                    'quantity' => 4,
+                    'sum' => 73.44,
                     'tax' => 'vat20',
                 ],
                 '100568_1' => [
-                    'price' => 28.27,
-                    'name' => 'ItbDrgQm',
-                    'quantity' => 2,
-                    'sum' => 56.54,
+                    'price' => 19.88,
+                    'name' => 'janvE9Ay',
+                    'quantity' => 19,
+                    'sum' => 377.72,
                     'tax' => 'vat20',
                 ],
                 '100568_2' => [
-                    'price' => 28.28,
-                    'name' => 'ItbDrgQm',
-                    'quantity' => 18,
-                    'sum' => 509.04,
+                    'price' => 19.89,
+                    'name' => 'janvE9Ay',
+                    'quantity' => 1,
+                    'sum' => 19.89,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
@@ -1000,142 +1082,142 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'items' => [
                 '100569_1' => [
                     'price' => 18.31,
-                    'name' => 'v24x4rdR',
-                    'quantity' => 9,
-                    'sum' => 164.79,
+                    'name' => 'aQQcXkLB',
+                    'quantity' => 20,
+                    'sum' => 366.2,
                     'tax' => 'vat20',
                 ],
                 '100569_2' => [
                     'price' => 18.32,
-                    'name' => 'v24x4rdR',
-                    'quantity' => 31,
-                    'sum' => 567.92,
+                    'name' => 'aQQcXkLB',
+                    'quantity' => 20,
+                    'sum' => 366.4,
                     'tax' => 'vat20',
                 ],
                 '100570_1' => [
-                    'price' => 18.31,
-                    'name' => 'fbt8jFkP',
-                    'quantity' => 7,
-                    'sum' => 128.17,
+                    'price' => 18.3,
+                    'name' => 'pP4cXE4c',
+                    'quantity' => 9,
+                    'sum' => 164.7,
                     'tax' => 'vat20',
                 ],
                 '100570_2' => [
-                    'price' => 18.32,
-                    'name' => 'fbt8jFkP',
-                    'quantity' => 23,
-                    'sum' => 421.36,
+                    'price' => 18.31,
+                    'name' => 'pP4cXE4c',
+                    'quantity' => 21,
+                    'sum' => 384.51,
                     'tax' => 'vat20',
                 ],
                 '100571_1' => [
                     'price' => 14.75,
-                    'name' => '8x5ke1tD',
-                    'quantity' => 16,
-                    'sum' => 236,
+                    'name' => 'NeU6lQRX',
+                    'quantity' => 27,
+                    'sum' => 398.25,
                     'tax' => 'vat20',
                 ],
                 '100571_2' => [
                     'price' => 14.76,
-                    'name' => '8x5ke1tD',
-                    'quantity' => 24,
-                    'sum' => 354.24,
+                    'name' => 'NeU6lQRX',
+                    'quantity' => 13,
+                    'sum' => 191.88,
                     'tax' => 'vat20',
                 ],
                 '100572_1' => [
-                    'price' => 14.75,
-                    'name' => 'abCGLpaj',
-                    'quantity' => 20,
-                    'sum' => 295,
+                    'price' => 14.76,
+                    'name' => 'aHX2WaxX',
+                    'quantity' => 39,
+                    'sum' => 575.64,
                     'tax' => 'vat20',
                 ],
                 '100572_2' => [
-                    'price' => 14.76,
-                    'name' => 'abCGLpaj',
-                    'quantity' => 30,
-                    'sum' => 442.8,
+                    'price' => 14.77,
+                    'name' => 'aHX2WaxX',
+                    'quantity' => 11,
+                    'sum' => 162.47,
                     'tax' => 'vat20',
                 ],
                 '100573_1' => [
                     'price' => 18.31,
-                    'name' => 'X7sQC7ys',
-                    'quantity' => 7,
-                    'sum' => 128.17,
-                    'tax' => 'vat20',
-                ],
-                '100573_2' => [
-                    'price' => 18.32,
-                    'name' => 'X7sQC7ys',
-                    'quantity' => 23,
-                    'sum' => 421.36,
-                    'tax' => 'vat20',
-                ],
-                '100574_1' => [
-                    'price' => 18.31,
-                    'name' => 'JIiAUxQW',
+                    'name' => 'fLjhVF9j',
                     'quantity' => 2,
                     'sum' => 36.62,
                     'tax' => 'vat20',
                 ],
-                '100574_2' => [
+                '100573_2' => [
                     'price' => 18.32,
-                    'name' => 'JIiAUxQW',
+                    'name' => 'fLjhVF9j',
+                    'quantity' => 28,
+                    'sum' => 512.96,
+                    'tax' => 'vat20',
+                ],
+                '100574_1' => [
+                    'price' => 18.26,
+                    'name' => 'xuu28xie',
                     'quantity' => 8,
-                    'sum' => 146.56,
+                    'sum' => 146.08,
+                    'tax' => 'vat20',
+                ],
+                '100574_2' => [
+                    'price' => 18.27,
+                    'name' => 'xuu28xie',
+                    'quantity' => 2,
+                    'sum' => 36.54,
                     'tax' => 'vat20',
                 ],
                 '100575_1' => [
-                    'price' => 18.31,
-                    'name' => '7hqtKHLX',
-                    'quantity' => 12,
-                    'sum' => 219.72,
+                    'price' => 18.33,
+                    'name' => 'XjE4SYr8',
+                    'quantity' => 16,
+                    'sum' => 293.28,
                     'tax' => 'vat20',
                 ],
                 '100575_2' => [
-                    'price' => 18.32,
-                    'name' => '7hqtKHLX',
-                    'quantity' => 38,
-                    'sum' => 696.16,
+                    'price' => 18.34,
+                    'name' => 'XjE4SYr8',
+                    'quantity' => 34,
+                    'sum' => 623.56,
                     'tax' => 'vat20',
                 ],
                 '100576_1' => [
-                    'price' => 16.79,
-                    'name' => '4IfUtXvp',
-                    'quantity' => 9,
-                    'sum' => 151.11,
+                    'price' => 16.74,
+                    'name' => 'ZW6A11yF',
+                    'quantity' => 1,
+                    'sum' => 16.74,
                     'tax' => 'vat20',
                 ],
                 '100576_2' => [
-                    'price' => 16.8,
-                    'name' => '4IfUtXvp',
-                    'quantity' => 1,
-                    'sum' => 16.8,
+                    'price' => 16.75,
+                    'name' => 'ZW6A11yF',
+                    'quantity' => 9,
+                    'sum' => 150.75,
                     'tax' => 'vat20',
                 ],
                 '100577_1' => [
                     'price' => 18.31,
-                    'name' => 'cEdxQcNd',
-                    'quantity' => 5,
-                    'sum' => 91.55,
+                    'name' => 'UOn0RlkO',
+                    'quantity' => 1,
+                    'sum' => 18.31,
                     'tax' => 'vat20',
                 ],
                 '100577_2' => [
                     'price' => 18.32,
-                    'name' => 'cEdxQcNd',
-                    'quantity' => 15,
-                    'sum' => 274.8,
+                    'name' => 'UOn0RlkO',
+                    'quantity' => 19,
+                    'sum' => 348.08,
                     'tax' => 'vat20',
                 ],
                 '100578_1' => [
-                    'price' => 19.84,
-                    'name' => 'q5Lk3U2W',
-                    'quantity' => 12,
-                    'sum' => 238.08,
+                    'price' => 19.85,
+                    'name' => 'ZFV54mXu',
+                    'quantity' => 16,
+                    'sum' => 317.6,
                     'tax' => 'vat20',
                 ],
                 '100578_2' => [
-                    'price' => 19.85,
-                    'name' => 'q5Lk3U2W',
-                    'quantity' => 8,
-                    'sum' => 158.8,
+                    'price' => 19.86,
+                    'name' => 'ZFV54mXu',
+                    'quantity' => 4,
+                    'sum' => 79.44,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
@@ -1152,25 +1234,18 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'sum' => 7989.99,
             'origGrandTotal' => 7989.99,
             'items' => [
-                '100579_1' => [
-                    'price' => 30.75,
-                    'name' => '4kiTDqhU',
-                    'quantity' => 56,
-                    'sum' => 1722,
-                    'tax' => 'vat20',
-                ],
-                '100579_2' => [
-                    'price' => 30.76,
-                    'name' => '4kiTDqhU',
-                    'quantity' => 44,
-                    'sum' => 1353.44,
-                    'tax' => 'vat20',
-                ],
                 0 => [
-                    'price' => 4914.55,
-                    'name' => 'PqbrzrVx',
+                    'price' => 0,
+                    'name' => 'BNjkAE3U',
+                    'quantity' => 100,
+                    'sum' => 0,
+                    'tax' => 'vat20',
+                ],
+                1 => [
+                    'price' => 7989.99,
+                    'name' => 'hUSwdaHQ',
                     'quantity' => 1,
-                    'sum' => 4914.55,
+                    'sum' => 7989.99,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
@@ -1259,35 +1334,35 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'items' => [
                 '100586_1' => [
                     'price' => 1144.58,
-                    'name' => 'gDghRkvg',
+                    'name' => 'Lf7ji4Ms',
                     'quantity' => 2,
                     'sum' => 2289.16,
                     'tax' => 'vat20',
                 ],
                 '100586_2' => [
                     'price' => 1144.57,
-                    'name' => 'gDghRkvg',
+                    'name' => 'Lf7ji4Ms',
                     'quantity' => 3,
                     'sum' => 3433.71,
                     'tax' => 'vat20',
                 ],
                 '100587_1' => [
                     'price' => 2801.86,
-                    'name' => '9WgjLFHY',
+                    'name' => 'RD57qiHD',
                     'quantity' => 2,
                     'sum' => 5603.72,
                     'tax' => 'vat20',
                 ],
                 '100587_2' => [
                     'price' => 2801.85,
-                    'name' => '9WgjLFHY',
+                    'name' => 'RD57qiHD',
                     'quantity' => 1,
                     'sum' => 2801.85,
                     'tax' => 'vat20',
                 ],
                 0 => [
                     'price' => 543.16,
-                    'name' => '5b3e0rfX',
+                    'name' => 'L7vuod9b',
                     'quantity' => 1,
                     'sum' => 543.16,
                     'tax' => 'vat20',
@@ -1307,189 +1382,105 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'origGrandTotal' => 17431.01,
             'items' => [
                 100596 => [
-                    'price' => 6350.7,
+                    'price' => 1,
                     'quantity' => 1,
-                    'sum' => 6350.7,
+                    'sum' => 1,
                     'tax' => 'vat20',
                 ],
                 '100597_1' => [
-                    'price' => 18.43,
-                    'quantity' => 14,
-                    'sum' => 258.02,
+                    'price' => 29.01,
+                    'quantity' => 1,
+                    'sum' => 29.01,
                     'tax' => 'vat20',
                 ],
                 '100597_2' => [
-                    'price' => 18.44,
-                    'quantity' => 16,
-                    'sum' => 295.04,
+                    'price' => 29,
+                    'quantity' => 29,
+                    'sum' => 841,
                     'tax' => 'vat20',
                 ],
-                '100598_1' => [
-                    'price' => 23.52,
-                    'quantity' => 36,
-                    'sum' => 846.72,
+                100598 => [
+                    'price' => 37,
+                    'quantity' => 40,
+                    'sum' => 1480,
                     'tax' => 'vat20',
                 ],
-                '100598_2' => [
-                    'price' => 23.53,
-                    'quantity' => 4,
-                    'sum' => 94.12,
+                100599 => [
+                    'price' => 37,
+                    'quantity' => 40,
+                    'sum' => 1480,
                     'tax' => 'vat20',
                 ],
-                '100599_1' => [
-                    'price' => 23.52,
-                    'quantity' => 36,
-                    'sum' => 846.72,
+                100600 => [
+                    'price' => 37,
+                    'quantity' => 40,
+                    'sum' => 1480,
                     'tax' => 'vat20',
                 ],
-                '100599_2' => [
-                    'price' => 23.53,
-                    'quantity' => 4,
-                    'sum' => 94.12,
+                100601 => [
+                    'price' => 37,
+                    'quantity' => 40,
+                    'sum' => 1480,
                     'tax' => 'vat20',
                 ],
-                '100600_1' => [
-                    'price' => 23.52,
-                    'quantity' => 36,
-                    'sum' => 846.72,
+                100602 => [
+                    'price' => 37,
+                    'quantity' => 40,
+                    'sum' => 1480,
                     'tax' => 'vat20',
                 ],
-                '100600_2' => [
-                    'price' => 23.53,
-                    'quantity' => 4,
-                    'sum' => 94.12,
+                100603 => [
+                    'price' => 36,
+                    'quantity' => 10,
+                    'sum' => 360,
                     'tax' => 'vat20',
                 ],
-                '100601_1' => [
-                    'price' => 23.52,
-                    'quantity' => 36,
-                    'sum' => 846.72,
+                100604 => [
+                    'price' => 29,
+                    'quantity' => 60,
+                    'sum' => 1740,
                     'tax' => 'vat20',
                 ],
-                '100601_2' => [
-                    'price' => 23.53,
-                    'quantity' => 4,
-                    'sum' => 94.12,
+                100605 => [
+                    'price' => 29,
+                    'quantity' => 80,
+                    'sum' => 2320,
                     'tax' => 'vat20',
                 ],
-                '100602_1' => [
-                    'price' => 23.52,
-                    'quantity' => 36,
-                    'sum' => 846.72,
+                100606 => [
+                    'price' => 33,
+                    'quantity' => 30,
+                    'sum' => 990,
                     'tax' => 'vat20',
                 ],
-                '100602_2' => [
-                    'price' => 23.53,
-                    'quantity' => 4,
-                    'sum' => 94.12,
+                100607 => [
+                    'price' => 33,
+                    'quantity' => 20,
+                    'sum' => 660,
                     'tax' => 'vat20',
                 ],
-                '100603_1' => [
-                    'price' => 22.88,
-                    'quantity' => 5,
-                    'sum' => 114.4,
+                100608 => [
+                    'price' => 33,
+                    'quantity' => 10,
+                    'sum' => 330,
                     'tax' => 'vat20',
                 ],
-                '100603_2' => [
-                    'price' => 22.89,
-                    'quantity' => 5,
-                    'sum' => 114.45,
+                100609 => [
+                    'price' => 46,
+                    'quantity' => 20,
+                    'sum' => 920,
                     'tax' => 'vat20',
                 ],
-                '100604_1' => [
-                    'price' => 18.43,
-                    'quantity' => 27,
-                    'sum' => 497.61,
+                100610 => [
+                    'price' => 46,
+                    'quantity' => 20,
+                    'sum' => 920,
                     'tax' => 'vat20',
                 ],
-                '100604_2' => [
-                    'price' => 18.44,
-                    'quantity' => 33,
-                    'sum' => 608.52,
-                    'tax' => 'vat20',
-                ],
-                '100605_1' => [
-                    'price' => 18.43,
-                    'quantity' => 37,
-                    'sum' => 681.91,
-                    'tax' => 'vat20',
-                ],
-                '100605_2' => [
-                    'price' => 18.44,
-                    'quantity' => 43,
-                    'sum' => 792.92,
-                    'tax' => 'vat20',
-                ],
-                '100606_1' => [
-                    'price' => 20.97,
-                    'quantity' => 5,
-                    'sum' => 104.85,
-                    'tax' => 'vat20',
-                ],
-                '100606_2' => [
-                    'price' => 20.98,
-                    'quantity' => 25,
-                    'sum' => 524.5,
-                    'tax' => 'vat20',
-                ],
-                '100607_1' => [
-                    'price' => 20.97,
-                    'quantity' => 4,
-                    'sum' => 83.88,
-                    'tax' => 'vat20',
-                ],
-                '100607_2' => [
-                    'price' => 20.98,
-                    'quantity' => 16,
-                    'sum' => 335.68,
-                    'tax' => 'vat20',
-                ],
-                '100608_1' => [
-                    'price' => 20.97,
-                    'quantity' => 2,
-                    'sum' => 41.94,
-                    'tax' => 'vat20',
-                ],
-                '100608_2' => [
-                    'price' => 20.98,
-                    'quantity' => 8,
-                    'sum' => 167.84,
-                    'tax' => 'vat20',
-                ],
-                '100609_1' => [
-                    'price' => 29.24,
-                    'quantity' => 15,
-                    'sum' => 438.6,
-                    'tax' => 'vat20',
-                ],
-                '100609_2' => [
-                    'price' => 29.25,
-                    'quantity' => 5,
-                    'sum' => 146.25,
-                    'tax' => 'vat20',
-                ],
-                '100610_1' => [
-                    'price' => 29.24,
-                    'quantity' => 15,
-                    'sum' => 438.6,
-                    'tax' => 'vat20',
-                ],
-                '100610_2' => [
-                    'price' => 29.25,
-                    'quantity' => 5,
-                    'sum' => 146.25,
-                    'tax' => 'vat20',
-                ],
-                '100611_1' => [
-                    'price' => 29.24,
-                    'quantity' => 15,
-                    'sum' => 438.6,
-                    'tax' => 'vat20',
-                ],
-                '100611_2' => [
-                    'price' => 29.25,
-                    'quantity' => 5,
-                    'sum' => 146.25,
+                100611 => [
+                    'price' => 46,
+                    'quantity' => 20,
+                    'sum' => 920,
                     'tax' => 'vat20',
                 ],
                 100612 => [
@@ -1586,178 +1577,94 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
             'origGrandTotal' => 19830.0,
             'items' => [
                 100610 => [
-                    'price' => 6643.26,
+                    'price' => 0.0,
                     'name' => 'Product 1',
-                    'quantity' => 1,
-                    'sum' => 6643.26,
+                    'quantity' => 1.0,
+                    'sum' => 0.0,
                     'tax' => 'vat20',
                 ],
-                '100611_1' => [
-                    'price' => 65.83,
+                100611 => [
+                    'price' => 99.0,
                     'name' => 'Product 2',
-                    'quantity' => 6,
-                    'sum' => 394.98,
+                    'quantity' => 10.0,
+                    'sum' => 990.0,
                     'tax' => 'vat20',
                 ],
-                '100611_2' => [
-                    'price' => 65.84,
-                    'name' => 'Product 2',
-                    'quantity' => 4,
-                    'sum' => 263.36,
-                    'tax' => 'vat20',
-                ],
-                '100612_1' => [
-                    'price' => 27.26,
+                100612 => [
+                    'price' => 41.0,
                     'name' => 'Product 3',
-                    'quantity' => 5,
-                    'sum' => 136.3,
+                    'quantity' => 10.0,
+                    'sum' => 410.0,
                     'tax' => 'vat20',
                 ],
-                '100612_2' => [
-                    'price' => 27.27,
-                    'name' => 'Product 3',
-                    'quantity' => 5,
-                    'sum' => 136.35,
-                    'tax' => 'vat20',
-                ],
-                '100613_1' => [
-                    'price' => 27.26,
+                100613 => [
+                    'price' => 41.0,
                     'name' => 'Product 4',
-                    'quantity' => 11,
-                    'sum' => 299.86,
+                    'quantity' => 20.0,
+                    'sum' => 820.0,
                     'tax' => 'vat20',
                 ],
-                '100613_2' => [
-                    'price' => 27.27,
-                    'name' => 'Product 4',
-                    'quantity' => 9,
-                    'sum' => 245.43,
-                    'tax' => 'vat20',
-                ],
-                '100614_1' => [
-                    'price' => 20.61,
+                100614 => [
+                    'price' => 31.0,
                     'name' => 'Product 5',
-                    'quantity' => 27,
-                    'sum' => 556.47,
+                    'quantity' => 50.0,
+                    'sum' => 1550.0,
                     'tax' => 'vat20',
                 ],
-                '100614_2' => [
-                    'price' => 20.62,
-                    'name' => 'Product 5',
-                    'quantity' => 23,
-                    'sum' => 474.26,
-                    'tax' => 'vat20',
-                ],
-                '100615_1' => [
-                    'price' => 20.61,
+                100615 => [
+                    'price' => 31.0,
                     'name' => 'Product 6',
-                    'quantity' => 32,
-                    'sum' => 659.52,
+                    'quantity' => 60.0,
+                    'sum' => 1860.0,
                     'tax' => 'vat20',
                 ],
-                '100615_2' => [
-                    'price' => 20.62,
-                    'name' => 'Product 6',
-                    'quantity' => 28,
-                    'sum' => 577.36,
-                    'tax' => 'vat20',
-                ],
-                '100616_1' => [
-                    'price' => 27.26,
+                100616 => [
+                    'price' => 41.0,
                     'name' => 'Product 7',
-                    'quantity' => 27,
-                    'sum' => 736.02,
+                    'quantity' => 50.0,
+                    'sum' => 2050.0,
                     'tax' => 'vat20',
                 ],
-                '100616_2' => [
-                    'price' => 27.27,
-                    'name' => 'Product 7',
-                    'quantity' => 23,
-                    'sum' => 627.21,
-                    'tax' => 'vat20',
-                ],
-                '100617_1' => [
-                    'price' => 25.93,
+                100617 => [
+                    'price' => 39.0,
                     'name' => 'Product 8',
-                    'quantity' => 27,
-                    'sum' => 700.11,
+                    'quantity' => 50.0,
+                    'sum' => 1950.0,
                     'tax' => 'vat20',
                 ],
-                '100617_2' => [
-                    'price' => 25.94,
-                    'name' => 'Product 8',
-                    'quantity' => 23,
-                    'sum' => 596.62,
-                    'tax' => 'vat20',
-                ],
-                '100618_1' => [
-                    'price' => 27.26,
+                100618 => [
+                    'price' => 41.0,
                     'name' => 'Product 9',
-                    'quantity' => 27,
-                    'sum' => 736.02,
+                    'quantity' => 50.0,
+                    'sum' => 2050.0,
                     'tax' => 'vat20',
                 ],
-                '100618_2' => [
-                    'price' => 27.27,
-                    'name' => 'Product 9',
-                    'quantity' => 23,
-                    'sum' => 627.21,
-                    'tax' => 'vat20',
-                ],
-                '100619_1' => [
-                    'price' => 27.26,
+                100619 => [
+                    'price' => 41.0,
                     'name' => 'Product 10',
-                    'quantity' => 27,
-                    'sum' => 736.02,
+                    'quantity' => 50.0,
+                    'sum' => 2050.0,
                     'tax' => 'vat20',
                 ],
-                '100619_2' => [
-                    'price' => 27.27,
-                    'name' => 'Product 10',
-                    'quantity' => 23,
-                    'sum' => 627.21,
-                    'tax' => 'vat20',
-                ],
-                '100620_1' => [
-                    'price' => 27.26,
+                100620 => [
+                    'price' => 41.0,
                     'name' => 'Product 11',
-                    'quantity' => 27,
-                    'sum' => 736.02,
+                    'quantity' => 50.0,
+                    'sum' => 2050.0,
                     'tax' => 'vat20',
                 ],
-                '100620_2' => [
-                    'price' => 27.27,
-                    'name' => 'Product 11',
-                    'quantity' => 23,
-                    'sum' => 627.21,
-                    'tax' => 'vat20',
-                ],
-                '100621_1' => [
-                    'price' => 20.61,
+                100621 => [
+                    'price' => 31.0,
                     'name' => 'Product 12',
-                    'quantity' => 27,
-                    'sum' => 556.47,
+                    'quantity' => 50.0,
+                    'sum' => 1550.0,
                     'tax' => 'vat20',
                 ],
-                '100621_2' => [
-                    'price' => 20.62,
-                    'name' => 'Product 12',
-                    'quantity' => 23,
-                    'sum' => 474.26,
-                    'tax' => 'vat20',
-                ],
-                '100622_1' => [
-                    'price' => 33.24,
+                100622 => [
+                    'price' => 50.0,
                     'name' => 'Product 13',
-                    'quantity' => 3,
-                    'sum' => 99.72,
-                    'tax' => 'vat20',
-                ],
-                '100622_2' => [
-                    'price' => 33.25,
-                    'name' => 'Product 13',
-                    'quantity' => 47,
-                    'sum' => 1562.75,
+                    'quantity' => 50.0,
+                    'sum' => 2500.0,
                     'tax' => 'vat20',
                 ],
                 100623 => [
@@ -1794,9 +1701,9 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
                     'tax' => 'vat20',
                 ],
                 100626 => [
-                    'price' => 5153.66,
+                    'price' => 10307.32,
                     'quantity' => 1.0,
-                    'sum' => 5153.66,
+                    'sum' => 10307.32,
                     'tax' => 'vat20',
                 ],
                 100627 => [
@@ -1806,9 +1713,9 @@ class DiscountSpreadAndSplitTest extends DiscountGeneralTestCase
                     'tax' => 'vat20',
                 ],
                 100628 => [
-                    'price' => 5153.66,
+                    'price' => 0.0,
                     'quantity' => 1.0,
-                    'sum' => 5153.66,
+                    'sum' => 0.0,
                     'tax' => 'vat20',
                 ],
                 100629 => [
