+++ Test/Unit/DiscountHelper/WithoutCalcTest.php
@@ -6,10 +6,16 @@
  * @package Mygento_Base
  */
 
-namespace Mygento\Base\Test\Unit;
+namespace Mygento\Base\Test\Unit\DiscountHelper;
 
-class DiscountAffectsShippingTest extends DiscountGeneralTestCase
+class WithoutCalcTest extends GeneralTestCase
 {
+    protected function setUp(): void
+    {
+        $this->discountHelper = $this->getDiscountHelperInstance();
+        $this->discountHelper->setDoCalculation(false);
+    }
+
     /**
      * Attention! Order of items in array is important!
      * @dataProvider dataProviderOrdersForCheckCalculation
@@ -35,11 +41,14 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
         foreach ($recalcItems as $index => $recalcItem) {
             $this->assertEquals($recalcExpectedItems[$index]['price'], $recalcItem['price'], 'Price of item failed');
             $this->assertEquals($recalcExpectedItems[$index]['quantity'], $recalcItem['quantity']);
-            $this->assertEquals($recalcExpectedItems[$index]['sum'], $recalcItem['sum'], 'Sum of item failed');
+
+            $sumEqual = bccomp($recalcExpectedItems[$index]['sum'], $recalcItem['sum']);
+            $this->assertEquals($sumEqual, 0, 'Sum of item failed');
         }
     }
 
-    /**
+    /** Для этой группы тестов - если есть глобальная скидка, то мы все равно пытаемся её распределить между позициями
+     * @return mixed
      * @SuppressWarnings(PHPMD.ExcessiveMethodLength)
      */
     protected static function getExpected()
@@ -48,23 +57,23 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'sum' => 12069.30,
             'origGrandTotal' => 12069.30,
             'items' => [
-                152 => [
-                    'price' => 11691,
+                0 => [
+                    'price' => 1,
                     'quantity' => 1,
                     'sum' => 11691,
                 ],
-                153 => [
-                    'price' => 378.30,
+                1 => [
+                    'price' => 1,
                     'quantity' => 1,
                     'sum' => 378.30,
                 ],
-                154 => [
-                    'price' => 0,
+                2 => [
+                    'price' => 1,
                     'quantity' => 1,
                     'sum' => 0,
                 ],
                 'shipping' => [
-                    'price' => 0.00,
+                    'price' => 0,
                     'quantity' => 1,
                     'sum' => 0,
                 ],
@@ -72,47 +81,47 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
         ];
 
         $actualData[parent::TEST_CASE_NAME_2] = [
-            'sum' => 5086.17,
+            'sum' => 5086.19,
             'origGrandTotal' => 9373.19,
             'items' => [
                 152 => [
-                    'price' => 5054.4,
+                    'price' => 1,
                     'quantity' => 1,
                     'sum' => 5054.4,
                 ],
                 153 => [
-                    'price' => 10.59,
+                    'price' => 1,
                     'quantity' => 3,
-                    'sum' => 31.77,
+                    'sum' => 31.79,
                 ],
                 'shipping' => [
-                    'price' => 4287.02,
+                    'price' => 4287.00,
                     'quantity' => 1,
-                    'sum' => 4287.02,
+                    'sum' => 4287.00,
                 ],
             ],
         ];
 
         $actualData[parent::TEST_CASE_NAME_3] = [
-            'sum' => 5086.17,
+            'sum' => 5086.19,
             'origGrandTotal' => 5106.19,
             'items' => [
                 0 => [
-                    'price' => 5015.28,
+                    'price' => 1,
                     'quantity' => 1,
                     'sum' => 5015.28,
                     'tax' => 'vat20',
                 ],
                 1 => [
-                    'price' => 23.63,
+                    'price' => 1,
                     'quantity' => 3,
-                    'sum' => 70.89,
+                    'sum' => 70.91,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
-                    'price' => 20.02,
+                    'price' => 20,
                     'quantity' => 1,
-                    'sum' => 20.02,
+                    'sum' => 20,
                     'tax' => '',
                 ],
             ],
@@ -123,12 +132,12 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'origGrandTotal' => 5200.86,
             'items' => [
                 152 => [
-                    'price' => 500.41,
+                    'price' => 1,
                     'quantity' => 2,
                     'sum' => 1000.82,
                 ],
                 153 => [
-                    'price' => 1000.01,
+                    'price' => 1,
                     'quantity' => 4,
                     'sum' => 4000.04,
                 ],
@@ -141,111 +150,123 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
         ];
 
         $actualData[parent::TEST_CASE_NAME_5] = [
-            'sum' => 202.06,
+            'sum' => 202.1,
             'origGrandTotal' => 202.1,
             'items' => [
-                152 => [
-                    'price' => 33.66,
+                '152_1' => [
+                    'price' => 1,
                     'quantity' => 3,
-                    'sum' => 100.98,
+                    'sum' => 101,
                 ],
-                153 => [
-                    'price' => 25.27,
+                '153_1' => [
+                    'price' => 1,
                     'quantity' => 4,
-                    'sum' => 101.08,
+                    'sum' => 101.01,
                 ],
                 'shipping' => [
-                    'price' => 0.04,
+                    'price' => 0.0,
                     'quantity' => 1,
-                    'sum' => 0.04,
+                    'sum' => 0.00,
                 ],
             ],
         ];
 
         $actualData[parent::TEST_CASE_NAME_6] = [
-            'sum' => 702.06,
+            'sum' => 702.1,
             'origGrandTotal' => 702.1,
             'items' => [
-                152 => [
-                    'price' => 33.66,
+                0 => [
+                    'price' => 1,
                     'quantity' => 3,
-                    'sum' => 100.98,
+                    'sum' => 101,
+                    'tax' => 'vat20',
                 ],
-                153 => [
-                    'price' => 25.27,
+                1 => [
+                    'price' => 1,
                     'quantity' => 4,
-                    'sum' => 101.08,
+                    'sum' => 101.1,
+                    'tax' => 'vat20',
                 ],
-                154 => [
-                    'price' => 100,
+                2 => [
+                    'price' => 1,
                     'quantity' => 5,
                     'sum' => 500,
+                    'tax' => 'vat20',
                 ],
                 'shipping' => [
-                    'price' => 0.04,
+                    'price' => 0,
                     'quantity' => 1,
-                    'sum' => 0.04,
+                    'sum' => 0,
+                    'tax' => '',
                 ],
             ],
         ];
 
         $actualData[parent::TEST_CASE_NAME_7] = [
-            'sum' => 11691.0,
-            'origGrandTotal' => 11691.0,
+            'sum' => 11691,
+            'origGrandTotal' => 11691,
             'items' => [
-                152 => [
-                    'price' => 11691.0,
+                0 => [
+                    'price' => 1,
                     'quantity' => 1,
-                    'sum' => 11691.0,
+                    'sum' => 11691,
+                    'tax' => 'vat20',
                 ],
-                153 => [
-                    'price' => 0.0,
+                1 => [
+                    'price' => 1,
                     'quantity' => 1,
-                    'sum' => 0.0,
+                    'sum' => 0,
+                    'tax' => 'vat20',
                 ],
                 'shipping' => [
-                    'price' => 0.00,
+                    'price' => 0,
                     'quantity' => 1,
-                    'sum' => 0.00,
+                    'sum' => 0,
+                    'tax' => '',
                 ],
             ],
         ];
 
         $actualData[parent::TEST_CASE_NAME_8] = [
             'sum' => 11610.99,
-            'origGrandTotal' => 11611.0,
+            'origGrandTotal' => 11611,
             'items' => [
-                152 => [
+                0 => [
                     'price' => 11591.15,
                     'quantity' => 1,
                     'sum' => 11591.15,
+                    'tax' => 'vat20',
                 ],
-                153 => [
+                1 => [
                     'price' => 19.84,
                     'quantity' => 1,
                     'sum' => 19.84,
+                    'tax' => 'vat20',
                 ],
                 'shipping' => [
                     'price' => 0.01,
                     'quantity' => 1,
                     'sum' => 0.01,
+                    'tax' => '',
                 ],
             ],
         ];
 
         $actualData[parent::TEST_CASE_NAME_9] = [
-            'sum' => 12890.0,
-            'origGrandTotal' => 12890.0,
+            'sum' => 12890,
+            'origGrandTotal' => 12890,
             'items' => [
-                152 => [
-                    'price' => 12890.0,
+                0 => [
+                    'price' => 12890,
                     'quantity' => 1,
-                    'sum' => 12890.0,
+                    'sum' => 12890,
+                    'tax' => 'vat20',
                 ],
                 'shipping' => [
-                    'price' => 0.00,
+                    'price' => 0,
                     'quantity' => 1,
-                    'sum' => 0.00,
+                    'sum' => 0,
+                    'tax' => '',
                 ],
             ],
         ];
@@ -254,20 +275,23 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'sum' => 12909.98,
             'origGrandTotal' => 12909.99,
             'items' => [
-                152 => [
+                0 => [
                     'price' => 12890.14,
                     'quantity' => 1,
                     'sum' => 12890.14,
+                    'tax' => 'vat20',
                 ],
-                153 => [
+                1 => [
                     'price' => 19.84,
                     'quantity' => 1,
                     'sum' => 19.84,
+                    'tax' => 'vat20',
                 ],
                 'shipping' => [
                     'price' => 0.01,
                     'quantity' => 1,
                     'sum' => 0.01,
+                    'tax' => '',
                 ],
             ],
         ];
@@ -277,25 +301,25 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'origGrandTotal' => 32130.01,
             'items' => [
                 0 => [
-                    'price' => 19990,
+                    'price' => 1,
                     'quantity' => 1,
                     'sum' => 19990,
                     'tax' => 'vat20',
                 ],
                 1 => [
-                    'price' => 19,
+                    'price' => 1,
                     'quantity' => 500,
                     'sum' => 9500,
                     'tax' => 'vat20',
                 ],
                 2 => [
-                    'price' => 1000.01,
+                    'price' => 1,
                     'quantity' => 1,
                     'sum' => 1000.01,
                     'tax' => 'vat20',
                 ],
                 3 => [
-                    'price' => 410,
+                    'price' => 1,
                     'quantity' => 4,
                     'sum' => 1640,
                     'tax' => 'vat20',
@@ -315,83 +339,71 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'items' => [
                 0 => [
                     'price' => 7989.99,
-                    'name' => 'P9MNKYhl',
                     'quantity' => 1,
                     'sum' => 7989.99,
                     'tax' => 'vat20',
                 ],
                 1 => [
                     'price' => 18.35,
-                    'name' => '4ERY91mu',
                     'quantity' => 40,
                     'sum' => 734,
                     'tax' => 'vat20',
                 ],
                 2 => [
                     'price' => 18.35,
-                    'name' => 'KfeQ4b7b',
                     'quantity' => 30,
                     'sum' => 550.5,
                     'tax' => 'vat20',
                 ],
                 3 => [
                     'price' => 14.78,
-                    'name' => 't5A2Rxrh',
                     'quantity' => 40,
                     'sum' => 591.2,
                     'tax' => 'vat20',
                 ],
                 4 => [
                     'price' => 14.78,
-                    'name' => 'hK09CiPH',
                     'quantity' => 50,
                     'sum' => 739,
                     'tax' => 'vat20',
                 ],
                 5 => [
                     'price' => 18.35,
-                    'name' => 'tcL2igh1',
                     'quantity' => 30,
                     'sum' => 550.5,
                     'tax' => 'vat20',
                 ],
                 6 => [
                     'price' => 18.35,
-                    'name' => 'xeFWcmLR',
                     'quantity' => 10,
                     'sum' => 183.5,
                     'tax' => 'vat20',
                 ],
                 7 => [
                     'price' => 18.35,
-                    'name' => 'igEqwmIn',
                     'quantity' => 50,
                     'sum' => 917.5,
                     'tax' => 'vat20',
                 ],
                 8 => [
                     'price' => 16.82,
-                    'name' => 'eEIuB9Qa',
                     'quantity' => 10,
                     'sum' => 168.2,
                     'tax' => 'vat20',
                 ],
                 9 => [
                     'price' => 18.35,
-                    'name' => 'Xbkx2msA',
                     'quantity' => 20,
                     'sum' => 367,
                     'tax' => 'vat20',
                 ],
                 10 => [
                     'price' => 19.88,
-                    'name' => 'Frg6nmw6',
                     'quantity' => 20,
                     'sum' => 397.6,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
-                    'name' => '',
                     'price' => 1,
                     'quantity' => 1,
                     'sum' => 1,
@@ -406,76 +418,65 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'items' => [
                 0 => [
                     'price' => 18.35,
-                    'name' => 'MZq6NgUu',
                     'quantity' => 40,
                     'sum' => 734,
                     'tax' => 'vat20',
                 ],
                 1 => [
                     'price' => 18.35,
-                    'name' => 'rFGIjBMZ',
                     'quantity' => 30,
                     'sum' => 550.5,
                     'tax' => 'vat20',
                 ],
                 2 => [
                     'price' => 14.78,
-                    'name' => 'LjayXom2',
                     'quantity' => 40,
                     'sum' => 591.2,
                     'tax' => 'vat20',
                 ],
                 3 => [
                     'price' => 14.78,
-                    'name' => 'n2uBhwkF',
                     'quantity' => 50,
                     'sum' => 739,
                     'tax' => 'vat20',
                 ],
                 4 => [
                     'price' => 18.35,
-                    'name' => 'JFe1Ch42',
                     'quantity' => 30,
                     'sum' => 550.5,
                     'tax' => 'vat20',
                 ],
                 5 => [
                     'price' => 18.35,
-                    'name' => 'lA3FKtTZ',
                     'quantity' => 10,
                     'sum' => 183.5,
                     'tax' => 'vat20',
                 ],
                 6 => [
                     'price' => 18.35,
-                    'name' => 'XKrOuj4K',
                     'quantity' => 50,
                     'sum' => 917.5,
                     'tax' => 'vat20',
                 ],
                 7 => [
                     'price' => 16.82,
-                    'name' => 'Xwd89uKe',
                     'quantity' => 10,
                     'sum' => 168.2,
                     'tax' => 'vat20',
                 ],
                 8 => [
                     'price' => 18.35,
-                    'name' => 'Zf8D4wlJ',
                     'quantity' => 20,
                     'sum' => 367,
                     'tax' => 'vat20',
                 ],
                 9 => [
                     'price' => 19.88,
-                    'name' => 'IerjMQ0P',
                     'quantity' => 20,
                     'sum' => 397.6,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
-                    'name' => '',
                     'price' => 0.99,
                     'quantity' => 1,
                     'sum' => 0.99,
@@ -490,83 +491,71 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'items' => [
                 0 => [
                     'price' => 7990.01,
-                    'name' => '5pX9HUkK',
                     'quantity' => 1,
                     'sum' => 7990.01,
                     'tax' => 'vat20',
                 ],
                 1 => [
                     'price' => 18.35,
-                    'name' => 'RVauXqDg',
                     'quantity' => 40,
                     'sum' => 734,
                     'tax' => 'vat20',
                 ],
                 2 => [
                     'price' => 18.35,
-                    'name' => 'eXMVmtNM',
                     'quantity' => 30,
                     'sum' => 550.5,
                     'tax' => 'vat20',
                 ],
                 3 => [
                     'price' => 14.78,
-                    'name' => 'rKpM7xsp',
                     'quantity' => 40,
                     'sum' => 591.2,
                     'tax' => 'vat20',
                 ],
                 4 => [
                     'price' => 14.78,
-                    'name' => 'f3UEyJsu',
                     'quantity' => 50,
                     'sum' => 739,
                     'tax' => 'vat20',
                 ],
                 5 => [
                     'price' => 18.35,
-                    'name' => 'QRZPcPcn',
                     'quantity' => 30,
                     'sum' => 550.5,
                     'tax' => 'vat20',
                 ],
                 6 => [
                     'price' => 18.35,
-                    'name' => 'SjPkhLxi',
                     'quantity' => 10,
                     'sum' => 183.5,
                     'tax' => 'vat20',
                 ],
                 7 => [
                     'price' => 18.35,
-                    'name' => 'ogIuOKZy',
                     'quantity' => 50,
                     'sum' => 917.5,
                     'tax' => 'vat20',
                 ],
                 8 => [
                     'price' => 16.82,
-                    'name' => 'ShFH5gnF',
                     'quantity' => 10,
                     'sum' => 168.2,
                     'tax' => 'vat20',
                 ],
                 9 => [
                     'price' => 18.35,
-                    'name' => 'ZdKlAkHt',
                     'quantity' => 20,
                     'sum' => 367,
                     'tax' => 'vat20',
                 ],
                 10 => [
                     'price' => 19.88,
-                    'name' => 'QOrfYuvX',
                     'quantity' => 20,
                     'sum' => 397.6,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
-                    'name' => '',
                     'price' => 1,
                     'quantity' => 1,
                     'sum' => 1,
@@ -581,83 +570,71 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'items' => [
                 0 => [
                     'price' => 7989.96,
-                    'name' => 'Z7aMJRHe',
                     'quantity' => 1,
                     'sum' => 7989.96,
                     'tax' => 'vat20',
                 ],
                 1 => [
                     'price' => 18.35,
-                    'name' => 'ZGzlY3lG',
                     'quantity' => 40,
                     'sum' => 734,
                     'tax' => 'vat20',
                 ],
                 2 => [
                     'price' => 18.35,
-                    'name' => '0QhYCfOx',
                     'quantity' => 30,
                     'sum' => 550.5,
                     'tax' => 'vat20',
                 ],
                 3 => [
                     'price' => 14.78,
-                    'name' => 'Ddebej9A',
                     'quantity' => 40,
                     'sum' => 591.2,
                     'tax' => 'vat20',
                 ],
                 4 => [
                     'price' => 14.78,
-                    'name' => 'x3B4cOC0',
                     'quantity' => 50,
                     'sum' => 739,
                     'tax' => 'vat20',
                 ],
                 5 => [
                     'price' => 18.35,
-                    'name' => 'y8k9XsO1',
                     'quantity' => 30,
                     'sum' => 550.5,
                     'tax' => 'vat20',
                 ],
                 6 => [
                     'price' => 18.35,
-                    'name' => 'jKIoyDr1',
                     'quantity' => 10,
                     'sum' => 183.5,
                     'tax' => 'vat20',
                 ],
                 7 => [
                     'price' => 18.35,
-                    'name' => 'ScpiBhc7',
                     'quantity' => 50,
                     'sum' => 917.5,
                     'tax' => 'vat20',
                 ],
                 8 => [
                     'price' => 16.82,
-                    'name' => 'ZsULLCo2',
                     'quantity' => 10,
                     'sum' => 168.2,
                     'tax' => 'vat20',
                 ],
                 9 => [
                     'price' => 18.35,
-                    'name' => 'C2mcTUaA',
                     'quantity' => 20,
                     'sum' => 367,
                     'tax' => 'vat20',
                 ],
                 10 => [
                     'price' => 19.88,
-                    'name' => 'XyibNKuf',
                     'quantity' => 20,
                     'sum' => 397.6,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
-                    'name' => '',
                     'price' => 0.73,
                     'quantity' => 1,
                     'sum' => 0.73,
@@ -672,76 +649,65 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'items' => [
                 0 => [
                     'price' => 18.31,
-                    'name' => 'PYGIbulV',
                     'quantity' => 40,
                     'sum' => 732.4,
                     'tax' => 'vat20',
                 ],
                 1 => [
                     'price' => 18.3,
-                    'name' => 'Fhor0s1t',
                     'quantity' => 30,
                     'sum' => 549,
                     'tax' => 'vat20',
                 ],
                 2 => [
                     'price' => 14.75,
-                    'name' => '2QE7DEhC',
                     'quantity' => 40,
                     'sum' => 590,
                     'tax' => 'vat20',
                 ],
                 3 => [
                     'price' => 14.76,
-                    'name' => 'cuhpCjPM',
                     'quantity' => 50,
                     'sum' => 738,
                     'tax' => 'vat20',
                 ],
                 4 => [
                     'price' => 18.31,
-                    'name' => 'jYG8dyAQ',
                     'quantity' => 30,
                     'sum' => 549.3,
                     'tax' => 'vat20',
                 ],
                 5 => [
                     'price' => 18.26,
-                    'name' => 'Wc9zz8mX',
                     'quantity' => 10,
                     'sum' => 182.6,
                     'tax' => 'vat20',
                 ],
                 6 => [
                     'price' => 18.33,
-                    'name' => 'ShtjbsVl',
                     'quantity' => 50,
                     'sum' => 916.5,
                     'tax' => 'vat20',
                 ],
                 7 => [
                     'price' => 16.74,
-                    'name' => 'JJLF9Lim',
                     'quantity' => 10,
                     'sum' => 167.4,
                     'tax' => 'vat20',
                 ],
                 8 => [
                     'price' => 18.31,
-                    'name' => 'Exyvd5Gy',
                     'quantity' => 20,
                     'sum' => 366.2,
                     'tax' => 'vat20',
                 ],
                 9 => [
                     'price' => 19.85,
-                    'name' => 'YLrmTbGC',
                     'quantity' => 20,
                     'sum' => 397,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
-                    'name' => '',
                     'price' => 1.61,
                     'quantity' => 1,
                     'sum' => 1.61,
@@ -756,20 +722,17 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'items' => [
                 0 => [
                     'price' => 0,
-                    'name' => 'NGUlxstu',
                     'quantity' => 100,
                     'sum' => 0,
                     'tax' => 'vat20',
                 ],
                 1 => [
                     'price' => 7989.99,
-                    'name' => 'fMLjGnBE',
                     'quantity' => 1,
                     'sum' => 7989.99,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
-                    'name' => '',
                     'price' => 0,
                     'quantity' => 1,
                     'sum' => 0,
@@ -840,21 +803,21 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'items' => [
                 0 => [
                     'price' => 1144.58,
-                    'name' => 'YJmwecYY',
+                    'name' => 'gVNPUJl8',
                     'quantity' => 5,
                     'sum' => 5722.9,
                     'tax' => 'vat20',
                 ],
                 1 => [
                     'price' => 2801.86,
-                    'name' => 'lm8sDuAm',
+                    'name' => 'xFWbm8aX',
                     'quantity' => 3,
                     'sum' => 8405.58,
                     'tax' => 'vat20',
                 ],
                 2 => [
                     'price' => 543.17,
-                    'name' => 'Gs9i60km',
+                    'name' => 'dgZoOh0z',
                     'quantity' => 1,
                     'sum' => 543.17,
                     'tax' => 'vat20',
@@ -871,7 +834,7 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
         ];
 
         $actualData[parent::TEST_CASE_NAME_21] = [
-            'sum' => 17431.3,
+            'sum' => 17431.01,
             'origGrandTotal' => 17431.01,
             'items' => [
                 100596 => [
@@ -881,105 +844,105 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
                     'tax' => 'vat20',
                 ],
                 100597 => [
-                    'price' => 29.01,
+                    'price' => 1,
                     'quantity' => 30,
-                    'sum' => 870.3,
+                    'sum' => 870.01,
                     'tax' => 'vat20',
                 ],
                 100598 => [
-                    'price' => 37,
+                    'price' => 1,
                     'quantity' => 40,
                     'sum' => 1480,
                     'tax' => 'vat20',
                 ],
                 100599 => [
-                    'price' => 37,
+                    'price' => 1,
                     'quantity' => 40,
                     'sum' => 1480,
                     'tax' => 'vat20',
                 ],
                 100600 => [
-                    'price' => 37,
+                    'price' => 1,
                     'quantity' => 40,
                     'sum' => 1480,
                     'tax' => 'vat20',
                 ],
                 100601 => [
-                    'price' => 37,
+                    'price' => 1,
                     'quantity' => 40,
                     'sum' => 1480,
                     'tax' => 'vat20',
                 ],
                 100602 => [
-                    'price' => 37,
+                    'price' => 1,
                     'quantity' => 40,
                     'sum' => 1480,
                     'tax' => 'vat20',
                 ],
                 100603 => [
-                    'price' => 36,
+                    'price' => 1,
                     'quantity' => 10,
                     'sum' => 360,
                     'tax' => 'vat20',
                 ],
                 100604 => [
-                    'price' => 29,
+                    'price' => 1,
                     'quantity' => 60,
                     'sum' => 1740,
                     'tax' => 'vat20',
                 ],
                 100605 => [
-                    'price' => 29,
+                    'price' => 1,
                     'quantity' => 80,
                     'sum' => 2320,
                     'tax' => 'vat20',
                 ],
                 100606 => [
-                    'price' => 33,
+                    'price' => 1,
                     'quantity' => 30,
                     'sum' => 990,
                     'tax' => 'vat20',
                 ],
                 100607 => [
-                    'price' => 33,
+                    'price' => 1,
                     'quantity' => 20,
                     'sum' => 660,
                     'tax' => 'vat20',
                 ],
                 100608 => [
-                    'price' => 33,
+                    'price' => 1,
                     'quantity' => 10,
                     'sum' => 330,
                     'tax' => 'vat20',
                 ],
                 100609 => [
-                    'price' => 46,
+                    'price' => 1,
                     'quantity' => 20,
                     'sum' => 920,
                     'tax' => 'vat20',
                 ],
                 100610 => [
-                    'price' => 46,
+                    'price' => 1,
                     'quantity' => 20,
                     'sum' => 920,
                     'tax' => 'vat20',
                 ],
                 100611 => [
-                    'price' => 46,
+                    'price' => 1,
                     'quantity' => 20,
                     'sum' => 920,
                     'tax' => 'vat20',
                 ],
                 100612 => [
-                    'price' => 0,
+                    'price' => 1,
                     'quantity' => 4,
                     'sum' => 0,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
-                    'price' => -0.29,
+                    'price' => 0,
                     'quantity' => 1,
-                    'sum' => -0.29,
+                    'sum' => 0,
                     'tax' => '',
                 ],
             ],
@@ -990,19 +953,22 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'origGrandTotal' => 10.0,
             'items' => [
                 100605 => [
-                    'price' => 0.0,
+                    'price' => 1,
+                    'name' => 'MI1yi2wG',
                     'quantity' => 1.0,
                     'sum' => 0.0,
                     'tax' => 'vat20',
                 ],
                 100606 => [
-                    'price' => 0.0,
+                    'price' => 1,
+                    'name' => 'oAScgwsB',
                     'quantity' => 1.0,
                     'sum' => 0.0,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
-                    'price' => 10.0,
+                    'name' => '',
+                    'price' => 10,
                     'quantity' => 1.0,
                     'sum' => 10.0,
                     'tax' => '',
@@ -1015,13 +981,13 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'origGrandTotal' => 200.0,
             'items' => [
                 100607 => [
-                    'price' => 0.0,
+                    'price' => 1,
                     'quantity' => 1.0,
                     'sum' => 0.0,
                     'tax' => 'vat20',
                 ],
                 100608 => [
-                    'price' => 0.0,
+                    'price' => 1,
                     'quantity' => 1.0,
                     'sum' => 0.0,
                     'tax' => 'vat20',
@@ -1040,8 +1006,8 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'origGrandTotal' => 0.0,
             'items' => [
                 100609 => [
-                    'price' => 0.0,
-                    'name' => '4j3FAJT7',
+                    'price' => 1,
+                    'name' => 'sGl8Rksk',
                     'quantity' => 1.0,
                     'sum' => 0.0,
                     'tax' => 'vat20',
@@ -1057,113 +1023,113 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
         ];
 
         $actualData[parent::TEST_CASE_NAME_25] = [
-            'sum' => 19830.00,
-            'origGrandTotal' => 19830,
+            'sum' => '19830.00',
+            'origGrandTotal' => 19830.0,
             'items' => [
-                100501 => [
-                    'price' => 0,
+                100610 => [
+                    'price' => 1,
                     'name' => 'Product 1',
-                    'quantity' => 1,
-                    'sum' => 0,
+                    'quantity' => 1.0,
+                    'sum' => 0.0,
                     'tax' => 'vat20',
                 ],
-                100502 => [
-                    'price' => 99,
+                100611 => [
+                    'price' => 1,
                     'name' => 'Product 2',
-                    'quantity' => 10,
-                    'sum' => 990,
+                    'quantity' => 10.0,
+                    'sum' => 990.0,
                     'tax' => 'vat20',
                 ],
-                100503 => [
-                    'price' => 41,
+                100612 => [
+                    'price' => 1,
                     'name' => 'Product 3',
-                    'quantity' => 10,
-                    'sum' => 410,
+                    'quantity' => 10.0,
+                    'sum' => 410.0,
                     'tax' => 'vat20',
                 ],
-                100504 => [
-                    'price' => 41,
+                100613 => [
+                    'price' => 1,
                     'name' => 'Product 4',
-                    'quantity' => 20,
-                    'sum' => 820,
+                    'quantity' => 20.0,
+                    'sum' => 820.0,
                     'tax' => 'vat20',
                 ],
-                100505 => [
-                    'price' => 31,
+                100614 => [
+                    'price' => 1,
                     'name' => 'Product 5',
-                    'quantity' => 50,
-                    'sum' => 1550,
+                    'quantity' => 50.0,
+                    'sum' => 1550.0,
                     'tax' => 'vat20',
                 ],
-                100506 => [
-                    'price' => 31,
+                100615 => [
+                    'price' => 1,
                     'name' => 'Product 6',
-                    'quantity' => 60,
-                    'sum' => 1860,
+                    'quantity' => 60.0,
+                    'sum' => 1860.0,
                     'tax' => 'vat20',
                 ],
-                100507 => [
-                    'price' => 41,
+                100616 => [
+                    'price' => 1,
                     'name' => 'Product 7',
-                    'quantity' => 50,
-                    'sum' => 2050,
+                    'quantity' => 50.0,
+                    'sum' => 2050.0,
                     'tax' => 'vat20',
                 ],
-                100508 => [
-                    'price' => 39,
+                100617 => [
+                    'price' => 1,
                     'name' => 'Product 8',
-                    'quantity' => 50,
-                    'sum' => 1950,
+                    'quantity' => 50.0,
+                    'sum' => 1950.0,
                     'tax' => 'vat20',
                 ],
-                100509 => [
-                    'price' => 41,
+                100618 => [
+                    'price' => 1,
                     'name' => 'Product 9',
-                    'quantity' => 50,
-                    'sum' => 2050,
+                    'quantity' => 50.0,
+                    'sum' => 2050.0,
                     'tax' => 'vat20',
                 ],
-                100510 => [
-                    'price' => 41,
+                100619 => [
+                    'price' => 1,
                     'name' => 'Product 10',
-                    'quantity' => 50,
-                    'sum' => 2050,
+                    'quantity' => 50.0,
+                    'sum' => 2050.0,
                     'tax' => 'vat20',
                 ],
-                100511 => [
-                    'price' => 41,
+                100620 => [
+                    'price' => 1,
                     'name' => 'Product 11',
-                    'quantity' => 50,
-                    'sum' => 2050,
+                    'quantity' => 50.0,
+                    'sum' => 2050.0,
                     'tax' => 'vat20',
                 ],
-                100512 => [
-                    'price' => 31,
+                100621 => [
+                    'price' => 1,
                     'name' => 'Product 12',
-                    'quantity' => 50,
-                    'sum' => 1550,
+                    'quantity' => 50.0,
+                    'sum' => 1550.0,
                     'tax' => 'vat20',
                 ],
-                100513 => [
-                    'price' => 50,
+                100622 => [
+                    'price' => 1,
                     'name' => 'Product 13',
-                    'quantity' => 50,
-                    'sum' => 2500,
+                    'quantity' => 50.0,
+                    'sum' => 2500.0,
                     'tax' => 'vat20',
                 ],
-                100514 => [
-                    'price' => 0,
+                100623 => [
+                    'price' => 1,
                     'name' => 'Product 14',
-                    'quantity' => 4,
-                    'sum' => 0,
+                    'quantity' => 4.0,
+                    'sum' => 0.0,
                     'tax' => 'vat20',
                 ],
                 'shipping' => [
                     'name' => '',
-                    'price' => 0,
-                    'quantity' => 1,
-                    'sum' => 0,
-                    'tax' => 'vat20',
+                    'price' => 0.0,
+                    'quantity' => 1.0,
+                    'sum' => 0.0,
+                    'tax' => '',
                 ],
             ],
         ];
@@ -1224,7 +1190,7 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'origGrandTotal' => 100.0,
             'items' => [
                 100630 => [
-                    'price' => 0.0,
+                    'price' => 1,
                     'quantity' => 1.0,
                     'sum' => 0.0,
                     'tax' => 'vat20',
@@ -1243,7 +1209,7 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'origGrandTotal' => 1320.0,
             'items' => [
                 100633 => [
-                    'price' => 1200.0,
+                    'price' => 1.0,
                     'quantity' => 1.0,
                     'sum' => 1200.0,
                     'tax' => 'vat20',
@@ -1262,7 +1228,7 @@ class DiscountAffectsShippingTest extends DiscountGeneralTestCase
             'origGrandTotal' => 1314.0,
             'items' => [
                 100634 => [
-                    'price' => 1200.0,
+                    'price' => 1.0,
                     'quantity' => 1.0,
                     'sum' => 1200.0,
                     'tax' => 'vat20',
