+++ Test/framework/autoload.php
@@ -1,0 +1,29 @@
+<?php
+
+/**
+ * @author Mygento Team
+ * @copyright 2014-2021 Mygento (https://www.mygento.ru)
+ * @package Mygento_Base
+ */
+
+use Magento\Framework\App\Filesystem\DirectoryList;
+use Magento\Framework\Code\Generator\Io;
+use Magento\Framework\Filesystem\Driver\File;
+use Magento\Framework\TestFramework\Unit\Autoloader\ExtensionAttributesGenerator;
+use Magento\Framework\TestFramework\Unit\Autoloader\ExtensionAttributesInterfaceGenerator;
+use Magento\Framework\TestFramework\Unit\Autoloader\FactoryGenerator;
+use Magento\Framework\TestFramework\Unit\Autoloader\GeneratedClassesAutoloader;
+
+$generatorIo = new Io(
+    new File(),
+    TESTS_TEMP_DIR . '/' . DirectoryList::getDefaultConfig()[DirectoryList::GENERATED_CODE][DirectoryList::PATH]
+);
+$generatedCodeAutoloader = new GeneratedClassesAutoloader(
+    [
+        new ExtensionAttributesGenerator(),
+        new ExtensionAttributesInterfaceGenerator(),
+        new FactoryGenerator(),
+    ],
+    $generatorIo
+);
+spl_autoload_register([$generatedCodeAutoloader, 'load']);
